# backend/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "QuickFlow API",
    "description": "Single-user productivity API: tasks, habits, learning resources, plans, dashboard. Errors are RFC 7807 problem+json; validation errors list {field, message} in `errors`.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/learning-cards/{id}": {
      "get": {
        "tags": [
          "Learning resources"
        ],
        "summary": "Get a learning card",
        "operationId": "get_3",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Card, milestone or note not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/LearningCard"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/LearningCard"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Learning resources"
        ],
        "summary": "Update a learning card (title, description, status)",
        "operationId": "update_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LearningCardRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "Card, milestone or note not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/LearningCard"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/LearningCard"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Learning resources"
        ],
        "summary": "Remove a learning card with its milestones and notes",
        "operationId": "delete_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "404": {
            "description": "Card, milestone or note not found"
          },
          "400": {
            "description": "Validation failed"
          }
        }
      }
    },
...
    "/api/tasks/{id}/complete": {
      "post": {
        "tags": [
          "Tasks"
        ],
        "summary": "Mark a task completed (status DONE)",
        "operationId": "complete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Task not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/learning-cards": {
      "get": {
        "tags": [
          "Learning resources"
        ],
        "summary": "List learning cards with milestones and notes",
        "operationId": "list_2",
        "responses": {
          "404": {
            "description": "Card, milestone or note not found",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LearningCard"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LearningCard"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Learning resources"
        ],
        "summary": "Add a learning card",
        "operationId": "create_2",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LearningCardRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "Card, milestone or note not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/LearningCard"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/LearningCard"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/LearningCard"
                }
              }
            }
          }
        }
      }
    },
    "/api/learning-cards/{id}/notes": {
      "post": {
        "tags": [
          "Learning resources"
        ],
        "summary": "Add a note; returns the card",
        "operationId": "addNote",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NoteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Added",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/LearningCard"
                }
              }
            }
          },
          "404": {
            "description": "Card, milestone or note not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/LearningCard"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/LearningCard"
                }
              }
            }
          }
        }
      }
    },
    "/api/learning-cards/{id}/milestones": {
      "post": {
        "tags": [
          "Learning resources"
        ],
        "summary": "Add a milestone; returns the card",
        "operationId": "addMilestone",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MilestoneRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Added",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/LearningCard"
                }
              }
            }
          },
          "404": {
            "description": "Card, milestone or note not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/LearningCard"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/LearningCard"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/habits/{id}/completions": {
      "get": {
        "tags": [
          "Habits"
        ],
        "summary": "List completions, newest first",
        "operationId": "completions",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Habit not found",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HabitCompletion"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HabitCompletion"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Habits"
        ],
        "summary": "Mark complete for a date (default today); returns the habit with updated progress",
        "operationId": "complete_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompletionRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Recorded",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Habit"
                }
              }
            }
          },
          "404": {
            "description": "Habit not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Habit"
                }
              }
            }
          },
          "409": {
            "description": "Already completed for that date, or habit inactive",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Habit"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Habit"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/learning-cards/{id}/milestones/{milestoneId}": {
      "delete": {
        "tags": [
          "Learning resources"
        ],
        "summary": "Remove a milestone; returns the card",
        "operationId": "removeMilestone",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "milestoneId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Card, milestone or note not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/LearningCard"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/LearningCard"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Learning resources"
        ],
        "summary": "Complete/reopen, rename or re-date a milestone; returns the card",
        "operationId": "updateMilestone",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "milestoneId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MilestoneUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "Card, milestone or note not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/LearningCard"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/LearningCard"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/learning-cards/{id}/notes/{noteId}": {
      "delete": {
        "tags": [
          "Learning resources"
        ],
        "summary": "Remove a note; returns the card",
        "operationId": "removeNote",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "noteId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Card, milestone or note not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/LearningCard"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/LearningCard"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Task": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "TODO",
              "IN_PROGRESS",
              "DONE"
            ]
          },
          "priority": {
            "type": "string",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH"
            ]
          },
          "dueDate": {
            "type": "string",
            "format": "date"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "completedAt": {
            "type": "string",
            "format": "date-time"
          },
          "archived": {
            "type": "boolean"
          },
          "overdue": {
            "type": "boolean",
            "description": "Due before today, not done, not archived"
          }
        }
      },
...
      "LearningCardRequest": {
        "required": [
          "title"
        ],
        "type": "object",
        "properties": {
          "title": {
            "maxLength": 200,
            "minLength": 0,
            "type": "string"
          },
          "description": {
            "maxLength": 2000,
            "minLength": 0,
            "type": "string",
            "description": "Description or source: link, book, course name"
          },
          "status": {
            "type": "string",
            "enum": [
              "NOT_STARTED",
              "IN_PROGRESS",
              "COMPLETED"
            ]
          }
        },
        "description": "Status defaults to NOT_STARTED"
      },
      "LearningCard": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "NOT_STARTED",
              "IN_PROGRESS",
              "COMPLETED"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "milestones": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Milestone"
            }
          },
          "notes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Note"
            }
          },
          "milestonesDone": {
            "type": "integer",
            "format": "int32"
          },
          "milestonesTotal": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "Milestone": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "done": {
            "type": "boolean"
          },
          "targetDate": {
            "type": "string",
            "format": "date"
          },
          "completedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Note": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "text": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Habit": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "frequency": {
            "type": "string",
            "enum": [
              "DAILY",
              "WEEKLY"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "active": {
            "type": "boolean"
          },
          "progress": {
            "$ref": "#/components/schemas/HabitProgress"
          }
        }
      },
      "HabitProgress": {
        "type": "object",
        "properties": {
          "completedToday": {
            "type": "boolean"
          },
          "completedThisPeriod": {
            "type": "boolean"
          },
          "currentStreak": {
            "type": "integer",
            "format": "int32"
          },
          "totalCompletions": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "NoteRequest": {
        "required": [
          "text"
        ],
        "type": "object",
        "properties": {
          "text": {
            "maxLength": 5000,
            "minLength": 0,
            "type": "string"
          }
        }
      },
      "MilestoneRequest": {
        "required": [
          "title"
        ],
        "type": "object",
        "properties": {
          "title": {
            "maxLength": 200,
            "minLength": 0,
            "type": "string"
          },
          "targetDate": {
            "type": "string",
            "format": "date"
          }
        }
      },
      "CompletionRequest": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "description": "Defaults to today; future dates are rejected",
            "format": "date"
          }
        }
      },
...
      "MilestoneUpdate": {
        "type": "object",
        "properties": {
          "title": {
            "maxLength": 200,
            "minLength": 1,
            "type": "string"
          },
          "done": {
            "type": "boolean"
          },
          "targetDate": {
            "type": "string",
            "format": "date"
          }
        },
        "description": "Only the fields present are changed"
      },
...
      "HabitCompletion": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "habitId": {
            "type": "integer",
            "format": "int64"
          },
          "completionDate": {
            "type": "string",
            "format": "date"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  }
}